another set of assorted signed/unsigned fixes, take 1001

+++ src/libgeodecomp/geometry/partitions/stripingpartition.h
@@ -63,7 +63,7 @@ public:
 
     Iterator operator[](unsigned pos) const
     {
-        Coord<DIM> cursor = dimensions.indexToCoord(pos) + origin;
+        Coord<DIM> cursor = dimensions.indexToCoord(static_cast<int>(pos)) + origin;
         return Iterator(origin, cursor, dimensions);
     }
